// fe/src/lib/api/generated/model/organizationScopeValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type OrganizationScopeValue = typeof OrganizationScopeValue[keyof typeof OrganizationScopeValue];


export const OrganizationScopeValue = {
  INSIDE_SCHOOL: 'INSIDE_SCHOOL',
  OUTSIDE_SCHOOL: 'OUTSIDE_SCHOOL',
} as const;